All migrations and models were finished

+++ database/migrations/2024_03_28_161733_create_bookings_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('bookings', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('room_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('room_id');
             $table->date('start_date');
             $table->date('end_date');
             $table->enum('status', ['pending', 'confirmed', 'cancelled']);
